test: keep the chat release-stream case in the existing plugins install suite
Adding a test file shifts the oxlint shard partition and the node test plan
counts, which surfaced as unrelated lint and plan failures. Fold the case into
the sibling suite instead so the file inventory stays unchanged.
